requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image

# pip install streamlit google-generativeai python-dotenv pillow
# pip install pdf2image
# pip install PyMuPDF